requirements.txt
requests
django
djangorestframework
tables
timeloop
phonenumbers
pandas
psycopg2
python-dotenv
git+https://github.com/Lotfull/django-postgres-copy.git
django-constance
django-picklefield
beautifulsoup4
